+++ R/getPerson.R
@@ -20,6 +20,6 @@
 #' } 
 getPerson <- function(){
     ret <- mirtCAT_post_internal(person=.MCE$person, design=.MCE$design,
-                                 has_answers=.MCE$test@has_answers)
+                                 has_answers=.MCE$test@has_answers, GUI=TRUE)
     ret
 }